requirements.txt
aldryn-apphooks-config==0.3.3
aldryn-boilerplates==0.7.5
aldryn-categories==1.0.4
aldryn-common==1.0.4
aldryn-forms==2.2.4
aldryn-newsblog==1.3.3
aldryn-people==1.2.2
aldryn-reversion==1.1.0
aldryn-translation-tools==0.2.1
Babel==2.4.0
backport-collections==0.1
certifi==2017.7.27.1
chardet==3.0.4
cmsplugin-filer==1.1.3
cssselect==1.0.1
cssutils==1.0.2
Django==1.8.18
django-absolute==0.3
django-appconf==1.0.2
django-appdata==0.1.6
django-classy-tags==0.8.0
django-cms==3.4.4
django-emailit==0.2.2
django-filer==1.2.8
django-formtools==2.0
django-mptt==0.8.7
django-parler==1.8
django-phonenumber-field==1.3.0
django-polymorphic==1.0.2
django-reversion==1.10.2
django-sekizai==0.10.0
django-simple-captcha==0.5.5
django-sizefield==0.9.1
django-sortedm2m==1.5.0
django-tablib==3.2
django-taggit==0.22.1
django-treebeard==4.1.2
djangocms-admin-style==1.2.7
djangocms-attributes-field==0.3.0
djangocms-column==1.7.0
djangocms-googlemap==1.1.1
djangocms-link==2.1.2
djangocms-snippet==1.9.2
djangocms-style==2.0.2
djangocms-text-ckeditor==3.5.0
djangocms-video==2.0.3
easy-thumbnails==2.3
et-xmlfile==1.0.1
html5lib==0.9999999
idna==2.5
jdcal==1.3
lxml==3.8.0
odfpy==1.3.5
olefile==0.44
openpyxl==2.4.8
phonenumbers==8.7.1
phonenumberslite==8.7.1
Pillow==4.2.1
premailer==3.1.1
python-dateutil==2.6.1
python-slugify==1.2.0
pytz==2017.2
PyYAML==3.12
requests==2.18.3
six==1.10.0
tablib==0.11.5
unicodecsv==0.14.1
Unidecode==0.4.21
urllib3==1.22
xlrd==1.0.0
xlwt==1.2.0
YURL==0.13
